ml-api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
scikit-learn==1.3.0
pandas==2.0.3
requests==2.31.0
python-whois==0.8.0
tldextract==5.1.1
urllib3==2.0.4 